+++ src/main/java/de/adorsys/keycloak/config/model/UserImport.java
@@ -19,11 +19,13 @@
 package de.adorsys.keycloak.config.model;
 
 import org.keycloak.representations.idm.UserRepresentation;
+import org.springframework.stereotype.Component;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+@Component
 public class UserImport extends UserRepresentation {
 
     @Override