+++ src/main/java/de/adorsys/keycloak/config/properties/ImportConfigProperties.java
@@ -25,7 +25,7 @@ import org.springframework.validation.annotation.Validated;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
-@Configuration
+@Configuration(proxyBeanMethods = false)
 @ConfigurationProperties(prefix = "import")
 @Validated
 public class ImportConfigProperties {